cleanup: split persistent browser runtime serialization
